revise ability and work data
